Extract iOS host-app socket plumbing into IOSHostChannel actor
Step #3 of the architectural roadmap. `IOSPreviewSession.swift`
previously mixed four concerns: orchestration, raw BSD socket plumbing
(listenFD/connectedFD/accept loop/DispatchSourceRead), JSON
request/response routing with continuations, and SwiftUI-domain
operations (`switchPreview`, `setTraits`, `fetchElements`).

Pull the bottom three (everything but orchestration) into
`Sources/PreviewsIOS/IOSHostChannel.swift` — a public actor that owns
the listen/connected FDs, the read source, the read buffer, and the
pending-response continuation map. Exposes `bindAndListen() -> Int`,
`awaitConnect(timeout:)`, `send(_:)`, `sendAndAwait(_:id:timeout:) -> Data`,
`close()`, plus an `isConnected` query.

`IOSPreviewSession.swift` shrinks 699 → 426 lines (-273); the channel
adds 347 lines, net +74 LOC for the abstraction. The session no longer
imports `Darwin` for socket APIs and holds a single `private let
channel = IOSHostChannel()` instead of five socket-state properties.
`stop()` becomes `async` (was synchronous-on-actor) — the only caller,
`IOSPreviewHandle.stop()`, already awaits it.

API shape changes for the channel:
- `send(_:)` takes `sending [String: Any]` so callers can pass dict
  literals across actor isolation without Swift 6 Sendable errors.
- `sendAndAwait(_:id:timeout:)` returns raw `Data` (was `[String: Any]`)
  for the same reason; `IOSPreviewSession.fetchElements` decodes the
  bytes itself.
- A new upfront `connectedFD >= 0` check in `sendAndAwait` fails fast
  with `connectionLost` when the peer has already disconnected —
  without it, the call would register a continuation only the timeout
  could resolve, racy under load.

New `IOSHostChannelTests` (5 tests) exercises the accept/disconnect
cleanup with a real loopback socket pair — the most error-prone bit
of the original socket layer:
- disconnect-then-send fails fast with connectionLost (early-check)
- pending sendAndAwait is failed by handleDisconnect when peer
  disconnects mid-flight (NOT via the timeout)
- close() is idempotent after a successful connect
- close() before/without bindAndListen is a no-op

Optional follow-up flagged in the plan: port `IOSHostChannel` from
BSD sockets to `NWConnection` so the project has one networking
abstraction (`DaemonClient` already uses NWConnection). Out of scope
for this PR; revisit if any host-app protocol changes warrant it.
