+++ Sources/PreviewsIOS/IOSPreviewSession.swift
@@ -27,13 +27,10 @@ public actor IOSPreviewSession {
     private let setupDylibPath: URL?
     public var currentTraits: PreviewTraits { traits }
 
-    // TCP socket state
-    private var listenFD: Int32 = -1
-    private var connectedFD: Int32 = -1
-    private var readSource: DispatchSourceRead?
-    private var readBuffer = Data()
-    // Data-typed continuations for Sendable compliance across task boundaries
-    private var pendingDataResponses: [String: CheckedContinuation<Data, Error>] = [:]
+    /// Transport to the iOS host app. Owns all socket state — bind,
+    /// accept, line-delimited JSON send/receive, lifecycle. Constructed
+    /// per-session and torn down by `stop()`.
+    private let channel = IOSHostChannel()
 
     public static let hostBundleID = "com.previewsmcp.host"
 
@@ -108,47 +105,7 @@ public actor IOSPreviewSession {
         stage("host app ready")
 
         // 4. Create TCP server on loopback, bind to ephemeral port
-        let serverFD = Darwin.socket(AF_INET, SOCK_STREAM, 0)
-        guard serverFD >= 0 else {
-            throw IOSPreviewSessionError.socketCreateFailed
-        }
-        var reuse: Int32 = 1
-        setsockopt(serverFD, SOL_SOCKET, SO_REUSEADDR, &reuse, socklen_t(MemoryLayout<Int32>.size))
-
-        var addr = sockaddr_in()
-        addr.sin_len = UInt8(MemoryLayout<sockaddr_in>.size)
-        addr.sin_family = sa_family_t(AF_INET)
-        addr.sin_port = 0  // ephemeral
-        addr.sin_addr.s_addr = inet_addr("127.0.0.1")
-
-        let bindResult = withUnsafePointer(to: &addr) { ptr in
-            ptr.withMemoryRebound(to: sockaddr.self, capacity: 1) { sockPtr in
-                Darwin.bind(serverFD, sockPtr, socklen_t(MemoryLayout<sockaddr_in>.size))
-            }
-        }
-        guard bindResult == 0 else {
-            Darwin.close(serverFD)
-            throw IOSPreviewSessionError.socketCreateFailed
-        }
-        guard Darwin.listen(serverFD, 1) == 0 else {
-            Darwin.close(serverFD)
-            throw IOSPreviewSessionError.socketCreateFailed
-        }
-
-        // Read the assigned port
-        var boundAddr = sockaddr_in()
-        var boundLen = socklen_t(MemoryLayout<sockaddr_in>.size)
-        let nameResult = withUnsafeMutablePointer(to: &boundAddr) { ptr in
-            ptr.withMemoryRebound(to: sockaddr.self, capacity: 1) { sockPtr in
-                getsockname(serverFD, sockPtr, &boundLen)
-            }
-        }
-        guard nameResult == 0 else {
-            Darwin.close(serverFD)
-            throw IOSPreviewSessionError.socketCreateFailed
-        }
-        let port = Int(UInt16(bigEndian: boundAddr.sin_port))
-        listenFD = serverFD
+        let port = try await channel.bindAndListen()
 
         // 5. Boot + install + launch with retry. The whole sequence is
         // retried because PR #141 CI has shown the iOS simulator
@@ -231,36 +188,16 @@ public actor IOSPreviewSession {
         // 6. Accept connection from host app (up to 10 seconds)
         stage("launched pid=\(pid); awaiting socket connection")
         await progress?.report(.connectingToApp, message: "Waiting for host app connection...")
-        try await acceptConnection(timeout: .seconds(10))
-        setupReadLoop()
+        try await channel.awaitConnect(timeout: .seconds(10))
         stage("connected; start complete")
 
         return pid
     }
 
-    /// Close the socket connection and clean up resources.
-    public func stop() {
-        // Fail any pending responses
-        for (_, continuation) in pendingDataResponses {
-            continuation.resume(throwing: IOSPreviewSessionError.connectionLost)
-        }
-        pendingDataResponses.removeAll()
-
-        // Cancel read source (its cancel handler closes connectedFD)
-        if let source = readSource {
-            source.cancel()
-            readSource = nil
-        } else if connectedFD >= 0 {
-            Darwin.close(connectedFD)
-        }
-        connectedFD = -1
-
-        if listenFD >= 0 {
-            Darwin.close(listenFD)
-            listenFD = -1
-        }
-
-        readBuffer.removeAll()
+    /// Close the socket connection and clean up resources. Idempotent —
+    /// `IOSHostChannel.close()` is safe to call when nothing was opened.
+    public func stop() async {
+        await channel.close()
     }
 
     // MARK: - Communication
@@ -268,7 +205,7 @@ public actor IOSPreviewSession {
     /// Recompile the preview and signal the host app to reload via socket.
     /// Waits for reloadAck to ensure SwiftUI environment has propagated.
     public func reload() async throws {
-        guard connectedFD >= 0 else {
+        guard await channel.isConnected else {
             throw IOSPreviewSessionError.notStarted
         }
 
@@ -287,7 +224,7 @@ public actor IOSPreviewSession {
         let compileResult = try await previewSession.compile()
 
         let requestID = UUID().uuidString
-        _ = try await sendAndAwait(
+        _ = try await channel.sendAndAwait(
             ["type": "reload", "id": requestID, "dylibPath": compileResult.dylibPath.path],
             id: requestID,
             timeout: .seconds(5)
@@ -298,7 +235,7 @@ public actor IOSPreviewSession {
     /// falls back to full recompile if structural changes are detected.
     @discardableResult
     public func handleSourceChange() async throws -> Bool {
-        guard connectedFD >= 0 else {
+        guard await channel.isConnected else {
             throw IOSPreviewSessionError.notStarted
         }
 
@@ -327,7 +264,7 @@ public actor IOSPreviewSession {
                 }
                 return entry
             }
-            sendMessage(["type": "literals", "changes": json])
+            await channel.send(["type": "literals", "changes": json])
             return true
         }
 
@@ -369,7 +306,7 @@ public actor IOSPreviewSession {
 
     /// Send a tap at the given point coordinates (in device points).
     public func sendTap(x: Double, y: Double) async throws {
-        sendMessage(["type": "touch", "action": "tap", "x": x, "y": y])
+        await channel.send(["type": "touch", "action": "tap", "x": x, "y": y])
         try await Task.sleep(for: .milliseconds(250))
     }
 
@@ -380,7 +317,7 @@ public actor IOSPreviewSession {
         duration: Double = 0.3,
         steps: Int = 10
     ) async throws {
-        sendMessage([
+        await channel.send([
             "type": "touch",
             "action": "swipe",
             "fromX": fromX, "fromY": fromY,
@@ -393,18 +330,20 @@ public actor IOSPreviewSession {
     /// Fetch the accessibility tree from the running preview.
     /// - Parameter filter: Filter mode: "all" (default), "interactable", or "labeled"
     public func fetchElements(filter: String = "all") async throws -> String {
-        guard connectedFD >= 0 else {
+        guard await channel.isConnected else {
             throw IOSPreviewSessionError.notStarted
         }
 
         let requestID = UUID().uuidString
-        let response = try await sendAndAwait(
+        let responseData = try await channel.sendAndAwait(
             ["type": "elements", "id": requestID, "filter": filter],
             id: requestID,
             timeout: .seconds(3)
         )
 
-        guard let tree = response["tree"] as? [String: Any] else {
+        guard let response = try? JSONSerialization.jsonObject(with: responseData) as? [String: Any],
+            let tree = response["tree"] as? [String: Any]
+        else {
             throw IOSPreviewSessionError.socketResponseTimeout("elementsResponse")
         }
 
@@ -423,195 +362,6 @@ public actor IOSPreviewSession {
     /// Capture a screenshot of the simulator.
     public func screenshot(jpegQuality: Double = 0.85) async throws -> Data {
         return try await simulatorManager.screenshotData(udid: deviceUDID, jpegQuality: jpegQuality)
-    }
-
-    // MARK: - Socket internals
-
-    /// Accept an incoming connection on the listen socket.
-    private func acceptConnection(timeout: Duration) async throws {
-        try await withThrowingTaskGroup(of: Int32.self) { group in
-            group.addTask { [listenFD] in
-                // The dispatch source is owned by the continuation closure
-                // but must also be cancellable from the task-cancellation
-                // handler when the timeout task throws. Without this path,
-                // `withCheckedThrowingContinuation` ignores cancellation
-                // and Task 1 stays suspended forever on the source — the
-                // 10s timer throws, the group waits for Task 1 to
-                // terminate, and the whole acceptConnection hangs until
-                // some upstream kills the caller (see iOS CI regression
-                // where a flaky host-app connection turned into a 20-min
-                // step timeout instead of a 10s clean failure).
-                let sourceBox = DispatchSourceBox()
-                return try await withTaskCancellationHandler {
-                    try await withCheckedThrowingContinuation {
-                        (cont: CheckedContinuation<Int32, Error>) in
-                        let source = DispatchSource.makeReadSource(
-                            fileDescriptor: listenFD, queue: .global())
-                        sourceBox.store(source)
-                        var resumed = false
-                        source.setEventHandler {
-                            source.cancel()
-                            guard !resumed else { return }
-                            resumed = true
-                            let clientFD = Darwin.accept(listenFD, nil, nil)
-                            if clientFD >= 0 {
-                                cont.resume(returning: clientFD)
-                            } else {
-                                cont.resume(throwing: IOSPreviewSessionError.socketAcceptFailed)
-                            }
-                        }
-                        source.setCancelHandler {
-                            guard !resumed else { return }
-                            resumed = true
-                            cont.resume(throwing: IOSPreviewSessionError.socketAcceptTimeout)
-                        }
-                        source.resume()
-                    }
-                } onCancel: {
-                    sourceBox.cancel()
-                }
-            }
-            group.addTask {
-                try await Task.sleep(for: timeout)
-                throw IOSPreviewSessionError.socketAcceptTimeout
-            }
-            let fd = try await group.next()!
-            group.cancelAll()
-            self.connectedFD = fd
-        }
-
-        // Close listen socket after successful accept — no longer needed
-        if listenFD >= 0 {
-            Darwin.close(listenFD)
-            listenFD = -1
-        }
-    }
-
-    /// Set up a read loop on the connected socket.
-    private func setupReadLoop() {
-        let fd = connectedFD
-        let source = DispatchSource.makeReadSource(fileDescriptor: fd, queue: .global())
-        source.setEventHandler { [weak self] in
-            var buf = [UInt8](repeating: 0, count: 8192)
-            let n = Darwin.read(fd, &buf, buf.count)
-            if n > 0 {
-                let data = Data(buf[0..<n])
-                if let self {
-                    Task { await self.processIncomingData(data) }
-                }
-            } else if n == 0 {
-                // EOF — host app disconnected
-                if let self {
-                    Task { await self.handleDisconnect() }
-                }
-            } else {
-                // read() error — treat as disconnect (ECONNRESET, etc.)
-                let err = errno
-                if err != EAGAIN && err != EWOULDBLOCK {
-                    if let self {
-                        Task { await self.handleDisconnect() }
-                    }
-                }
-            }
-        }
-        source.setCancelHandler {
-            Darwin.close(fd)
-        }
-        source.resume()
-        readSource = source
-    }
-
-    /// Process incoming data from the socket (actor-isolated).
-    private func processIncomingData(_ data: Data) {
-        readBuffer.append(data)
-
-        // Split on newlines and process complete messages
-        while let newlineIndex = readBuffer.firstIndex(of: 0x0A) {
-            let lineData = Data(readBuffer[readBuffer.startIndex..<newlineIndex])
-            readBuffer = Data(readBuffer[(newlineIndex + 1)...])
-
-            // Parse just enough to extract the id for routing
-            guard let message = try? JSONSerialization.jsonObject(with: lineData) as? [String: Any],
-                let id = message["id"] as? String
-            else {
-                continue
-            }
-
-            // Resume the waiting continuation with raw Data (Sendable-safe)
-            if let continuation = pendingDataResponses.removeValue(forKey: id) {
-                continuation.resume(returning: lineData)
-            }
-        }
-    }
-
-    /// Handle host app disconnect.
-    private func handleDisconnect() {
-        connectedFD = -1
-        for (_, continuation) in pendingDataResponses {
-            continuation.resume(throwing: IOSPreviewSessionError.connectionLost)
-        }
-        pendingDataResponses.removeAll()
-    }
-
-    /// Send a JSON message over the socket (fire-and-forget).
-    private func sendMessage(_ dict: [String: Any]) {
-        guard connectedFD >= 0,
-            var data = try? JSONSerialization.data(withJSONObject: dict)
-        else { return }
-        data.append(0x0A)  // newline delimiter
-        let fd = connectedFD
-        var writeFailed = false
-        data.withUnsafeBytes { buf in
-            guard let base = buf.baseAddress else { return }
-            var remaining = buf.count
-            var offset = 0
-            while remaining > 0 {
-                let n = Darwin.write(fd, base + offset, remaining)
-                if n <= 0 {
-                    writeFailed = true
-                    break
-                }
-                offset += n
-                remaining -= n
-            }
-        }
-        if writeFailed {
-            handleDisconnect()
-        }
-    }
-
-    /// Send a message and await a response with the matching id.
-    /// Races the response against a timeout. The continuation is registered on the actor,
-    /// and both the response path (processIncomingData) and the timeout path use
-    /// removeValue(forKey:) to ensure exactly one resumption.
-    private func sendAndAwait(
-        _ message: [String: Any], id: String, timeout: Duration
-    ) async throws -> [String: Any] {
-        sendMessage(message)
-
-        let responseData: Data = try await withCheckedThrowingContinuation { cont in
-            pendingDataResponses[id] = cont
-
-            // Timeout task: if no response arrives, fail the continuation.
-            // Uses removeValue to guarantee no double-resume with processIncomingData.
-            Task { [weak self] in
-                try? await Task.sleep(for: timeout)
-                guard let self else { return }
-                if let cont = await self.removePendingResponse(forKey: id) {
-                    cont.resume(throwing: IOSPreviewSessionError.socketResponseTimeout(id))
-                }
-            }
-        }
-
-        guard let dict = try? JSONSerialization.jsonObject(with: responseData) as? [String: Any] else {
-            throw IOSPreviewSessionError.socketResponseTimeout(id)
-        }
-        return dict
-    }
-
-    /// Remove and return a pending response continuation (actor-isolated, prevents double-resume).
-    private func removePendingResponse(forKey id: String) -> CheckedContinuation<Data, Error>? {
-        pendingDataResponses.removeValue(forKey: id)
     }
 
     // MARK: - Accessibility tree filtering
@@ -672,28 +422,4 @@ public enum IOSPreviewSessionError: Error, LocalizedError, CustomStringConvertib
     }
 
     public var errorDescription: String? { description }
-}
-
-/// Thread-safe holder for a `DispatchSourceRead` that needs to be
-/// cancelled from outside the continuation that owns it. Used by
-/// `acceptConnection` to bridge task-cancellation into dispatch-source
-/// lifecycle — DispatchSource is not Sendable, so a plain closure
-/// capture won't compile under Swift 6 strict concurrency.
-private final class DispatchSourceBox: @unchecked Sendable {
-    private let lock = NSLock()
-    private var source: DispatchSourceRead?
-
-    func store(_ s: DispatchSourceRead) {
-        lock.lock()
-        defer { lock.unlock() }
-        source = s
-    }
-
-    func cancel() {
-        lock.lock()
-        let s = source
-        source = nil
-        lock.unlock()
-        s?.cancel()
-    }
 }
